requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
firebase-admin==6.2.0
python-dotenv==1.0.0
gunicorn==20.1.0
